+++ src/main/java/com/DAO/UserDAOImpl.java
@@ -19,7 +19,7 @@ public class UserDAOImpl implements UserDAO{
 		try {
 			String query = "insert into usermaster(empn,username,qtrno,email,phone,password,usercreationdate,status) values(?,?,?,?,?,?,SYSDATE,'A')";
 			java.sql.PreparedStatement pstmt = conn.prepareStatement(query);
-			pstmt.setInt(1, us.getEmpn());
+			pstmt.setLong(1, us.getEmpn());
 			pstmt.setString(2, us.getUsername());
 			pstmt.setString(3, us.getQtrno());
 			pstmt.setString(4, us.getEmail());
@@ -43,21 +43,21 @@ public class UserDAOImpl implements UserDAO{
 
 
 	@Override
-	public User userLogin(int empn, String password) {
+	public User userLogin(long empn, String password) {
 	
 
 		User us = null;
 		try {
 			String query = "select * from usermaster where empn=? and password=?";
 			java.sql.PreparedStatement pstmt = conn.prepareStatement(query);
-			pstmt.setInt(1, empn);
+			pstmt.setLong(1, empn);
 			pstmt.setString(2, password);
 
 			java.sql.ResultSet rs = pstmt.executeQuery();
 
 			if (rs.next()) {
 				us = new User();  
-				us.setEmpn(rs.getInt("empn"));
+				us.setEmpn(rs.getLong("empn"));
 				us.setUsername(rs.getString("username"));
 				us.setQtrno(rs.getString("qtrno"));
 				us.setEmail(rs.getString("email"));
